requirements.txt
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
streamlit>=1.24.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0